requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
django-jazzmin==3.0.0
rest-framework-redesign==0.2.9
drf-material==0.3.3
django-extensions==3.2.3